Removing disable method from repository to add card token manager
